Added a long method to the Attributes to handle long-values for drawing ids.

+++ src/main/com/siigna/app/SiignaApplet.scala
@@ -75,7 +75,7 @@ class SiignaApplet extends Applet {
       val drawingId = getParameter("drawingId")
 
       if (drawingId != null) try {
-        val id = drawingId.toInt
+        val id = drawingId.toLong
         Drawing.setAttribute("id", id)
         Log.success("Applet: Found drawing: " + id)
       }
